GeoDjangoDocker/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2-binary==2.9.6
gunicorn==21.2.0